moved clip_fft out of core library to lighten depedencies
